refactor into InstPart
