+++ midimasterpiece/src/main/java/org/vibehistorian/midimasterpiece/midigenerator/ArpPart.java
@@ -6,29 +6,7 @@ import javax.xml.bind.annotation.XmlType;
 
 @XmlRootElement(name = "arpPart")
 @XmlType(propOrder = {})
-public class ArpPart {
-	private int instrument = 46;
-	
-	private int hitsPerPattern = 8;
-	private int chordSpan = 1;
-	private ChordSpanFill chordSpanFill = ChordSpanFill.ALL;
-	
-	private int pauseChance = 70;
-	private int exceptionChance = 5;
-	private boolean repeatableNotes = true;
-	private int patternRepeat = 2;
-	
-	private int transpose = 0;
-	
-	private int order = 1;
-	
-	private int patternSeed = 0;
-	private RhythmPattern pattern = RhythmPattern.RANDOM;
-	private int patternShift = 0;
-	
-	private int midiChannel = 10;
-	
-	private boolean muted = false;
+public class ArpPart extends InstPart {
 	
 	public ArpPart() {
 		
@@ -56,70 +34,6 @@ public class ArpPart {
 		this.chordSpanFill = fill;
 	}
 	
-	public int getInstrument() {
-		return instrument;
-	}
-	
-	public void setInstrument(int instrument) {
-		this.instrument = instrument;
-	}
-	
-	public int getHitsPerPattern() {
-		return hitsPerPattern;
-	}
-	
-	public void setHitsPerPattern(int hitsPerPattern) {
-		this.hitsPerPattern = hitsPerPattern;
-	}
-	
-	public int getChordSpan() {
-		return chordSpan;
-	}
-	
-	public void setChordSpan(int chordSpan) {
-		this.chordSpan = chordSpan;
-	}
-	
-	public int getPauseChance() {
-		return pauseChance;
-	}
-	
-	public void setPauseChance(int pauseChance) {
-		this.pauseChance = pauseChance;
-	}
-	
-	public int getExceptionChance() {
-		return exceptionChance;
-	}
-	
-	public void setExceptionChance(int exceptionChance) {
-		this.exceptionChance = exceptionChance;
-	}
-	
-	public boolean isRepeatableNotes() {
-		return repeatableNotes;
-	}
-	
-	public void setRepeatableNotes(boolean repeatableNotes) {
-		this.repeatableNotes = repeatableNotes;
-	}
-	
-	public int getPatternRepeat() {
-		return patternRepeat;
-	}
-	
-	public void setPatternRepeat(int patternRepeat) {
-		this.patternRepeat = patternRepeat;
-	}
-	
-	public int getTranspose() {
-		return transpose;
-	}
-	
-	public void setTranspose(int transpose) {
-		this.transpose = transpose;
-	}
-	
 	@XmlAttribute
 	public int getOrder() {
 		return order;
@@ -128,54 +42,5 @@ public class ArpPart {
 	public void setOrder(int order) {
 		this.order = order;
 	}
-	
-	public int getPatternSeed() {
-		return patternSeed;
-	}
-	
-	public void setPatternSeed(int patternSeed) {
-		this.patternSeed = patternSeed;
-	}
-	
-	public RhythmPattern getPattern() {
-		return pattern;
-	}
-	
-	public void setPattern(RhythmPattern pattern) {
-		this.pattern = pattern;
-	}
-	
-	public int getPatternShift() {
-		return patternShift;
-	}
-	
-	public void setPatternShift(int patternShift) {
-		this.patternShift = patternShift;
-	}
-	
-	public boolean isMuted() {
-		return muted;
-	}
-	
-	public void setMuted(boolean muted) {
-		this.muted = muted;
-	}
-	
-	public int getMidiChannel() {
-		return midiChannel;
-	}
-	
-	public void setMidiChannel(int midiChannel) {
-		this.midiChannel = midiChannel;
-	}
-	
-	public ChordSpanFill getChordSpanFill() {
-		return chordSpanFill;
-	}
-	
-	public void setChordSpanFill(ChordSpanFill cspanFill) {
-		this.chordSpanFill = cspanFill;
-	}
-	
 	
 }
